feat: 약관 동의 아키텍처 리팩토링 및 백엔드 API 연동 반영
